pyproject.toml

[build-system]
requires = ["poetry-core", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry]
name = "estimenergy"
description = "Estimate Energy Consumption"
authors = ["Lennard Beers <l.beers@outlook.de>"]
repository = "https://github.com/EuleMitKeule/EstimEnergy"
readme = "README.md"
packages = [{include = "estimenergy"}]
license = "MIT"
version = "0.0.0"

[tool.poetry.scripts]
api = "estimenergy.main:start"
drop = "estimenergy.db:drop_all"

[tool.poe.tasks]
run = { "script" = "estimenergy.main:start" }
test = "pytest --cov estimenergy --cov-report term-missing"
black = "black estimenergy tests custom_components/estimenergy_integration"
isort = "isort estimenergy tests custom_components/estimenergy_integration"
mypy = "mypy estimenergy tests custom_components/estimenergy_integration"
bandit = "bandit -r estimenergy custom_components/estimenergy_integration"
format.sequence = ["black", "isort", "mypy"]
format.ignore_fail = true
security.sequence = ["bandit"]
security.ignore_fail = true

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = {extras = ["standard"], version = "~0.20.0"}
fastapi = "~0.92.0"
fastapi-crudrouter = "~0.8.6"
aioesphomeapi = "~13.4.0"
tortoise_orm = "~0.19.3"
requests = "~2.28.2"
prometheus_client = "~0.16.0"
prometheus-fastapi-instrumentator = "~5.10.0"
python-dotenv = "~0.21.1"
PyYAML = "~6.0"
types-requests = "^2.28.11.17"
types-pyyaml = "^6.0.12.9"
pydantic-settings-yaml = "^0.1.1"
influxdb-client = {extras = ["async"], version = "^1.36.1"}
sqlmodel = "^0.0.8"
psycopg2-binary = "^2.9.6"

[tool.poetry.group.dev.dependencies]
pytest = "~7.2.1"
pytest-cov = "~4.0.0"
httpx = "~0.23.3"
PyYAML = "~6.0"
poetry = "~1.4.0"
pytest-asyncio = "^0.20.3"
pytest-mock = "^3.10.0"
freezegun = "^1.2.2"
black = "^23.3.0"
mypy = "^1.1.1"
isort = "^5.12.0"
ruff = "^0.0.260"
bandit = "^1.7.5"
pycodestyle = "^2.10.0"
pydocstyle = "^6.3.0"
pylint = "^2.17.2"

[tool.poetry.group.build.dependencies]
poetry-dynamic-versioning = "~0.21.4"
twine = "~4.0.2"

[tool.isort]
profile = "black"
force_sort_within_sections = true
forced_separate = [
    "tests",
]
combine_as_imports = true

[tool.mypy]
ignore_missing_imports = true
disable_error_code = "call-arg"

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "not-context-manager",
]